fix directory navigation
