+++ examples/src/main/kotlin/com/algolia/instantsearch/examples/directory/DirectoryAdapter.kt
@@ -1,14 +1,17 @@
 package com.algolia.instantsearch.examples.directory
 
+import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import com.algolia.instantsearch.android.highlighting.toSpannedString
 import com.algolia.instantsearch.android.inflate
 import com.algolia.instantsearch.core.hits.HitsView
 import com.algolia.instantsearch.examples.R
 
-class DirectoryAdapter : ListAdapter<DirectoryItem, DirectoryViewHolder>(diffUtil),
+class DirectoryAdapter : ListAdapter<DirectoryItem, DirectoryViewHolder>(DirectoryAdapter),
     HitsView<DirectoryItem> {
 
     private enum class ViewType {
@@ -44,21 +47,37 @@ class DirectoryAdapter : ListAdapter<DirectoryItem, DirectoryViewHolder>(diffUti
         submitList(hits)
     }
 
-    companion object {
+    companion object : DiffUtil.ItemCallback<DirectoryItem>() {
 
-        private val diffUtil = object : DiffUtil.ItemCallback<DirectoryItem>() {
+        override fun areItemsTheSame(oldItem: DirectoryItem, newItem: DirectoryItem): Boolean {
+            return oldItem is DirectoryItem.Item
+                && newItem is DirectoryItem.Item
+                && oldItem.hit.objectID == newItem.hit.objectID
+        }
 
-            override fun areItemsTheSame(oldItem: DirectoryItem, newItem: DirectoryItem): Boolean {
-                return if (oldItem is DirectoryItem.Item && newItem is DirectoryItem.Item) {
-                    return oldItem.hit.objectID == newItem.hit.objectID
-                } else false
-            }
+        override fun areContentsTheSame(oldItem: DirectoryItem, newItem: DirectoryItem): Boolean {
+            return oldItem == newItem
+        }
+    }
+}
 
-            override fun areContentsTheSame(
-                oldItem: DirectoryItem,
-                newItem: DirectoryItem
-            ): Boolean {
-                return oldItem == newItem
+sealed class DirectoryViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+
+    data class Header(val view: TextView) : DirectoryViewHolder(view) {
+
+        fun bind(item: DirectoryItem.Header) {
+            view.text = item.name
+        }
+    }
+
+    data class Item(val view: View) : DirectoryViewHolder(view) {
+
+        fun bind(item: DirectoryItem.Item) {
+            val text = item.hit.highlightedName?.toSpannedString() ?: item.hit.name
+            view.findViewById<TextView>(R.id.itemName).text = text
+
+            view.setOnClickListener {
+                view.context.navigateTo(item)
             }
         }
     }
