--- examples/src/main/kotlin/com/algolia/instantsearch/examples/directory/DirectoryHit.kt
@@ -1,21 +1,0 @@
-package com.algolia.instantsearch.examples.directory
-
-import com.algolia.instantsearch.highlighting.Highlightable
-import com.algolia.search.model.Attribute
-import com.algolia.search.model.ObjectID
-import com.algolia.search.model.indexing.Indexable
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonObject
-
-@Serializable
-data class DirectoryHit(
-    override val objectID: ObjectID,
-    val name: String,
-    val type: String,
-    val index: String? = null,
-    override val _highlightResult: JsonObject? = null
-) : Indexable, Highlightable {
-
-    val highlightedName
-        get() = getHighlight(Attribute("name"), preTag = "<b>", postTag = "</b>")
-}
