--- examples/src/main/kotlin/com/algolia/instantsearch/examples/directory/DirectoryViewHolder.kt
@@ -1,35 +1,0 @@
-package com.algolia.instantsearch.examples.directory
-
-import android.content.Intent
-import android.view.View
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.algolia.instantsearch.android.highlighting.toSpannedString
-import com.algolia.instantsearch.examples.R
-import com.algolia.search.serialize.KeyName
-
-sealed class DirectoryViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-
-    data class Header(val view: TextView) : DirectoryViewHolder(view) {
-
-        fun bind(item: DirectoryItem.Header) {
-            view.text = item.name
-        }
-    }
-
-    data class Item(val view: View) : DirectoryViewHolder(view) {
-
-        fun bind(item: DirectoryItem.Item) {
-            val text = item.hit.highlightedName?.toSpannedString() ?: item.hit.name
-            view.findViewById<TextView>(R.id.itemName).text = text
-
-            view.setOnClickListener {
-                val intent = Intent(view.context, guides.getValue(item.hit.objectID).java).apply {
-                    putExtra(KeyName, item.hit.name)
-                }
-
-                view.context.startActivity(intent)
-            }
-        }
-    }
-}
