+++ examples/src/main/kotlin/com/algolia/instantsearch/examples/showcase/compose/directory/ComposeDirectoryShowcase.kt
@@ -1,6 +1,5 @@
 package com.algolia.instantsearch.examples.showcase.compose.directory
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -10,15 +9,13 @@ import com.algolia.instantsearch.core.connection.ConnectionHandler
 import com.algolia.instantsearch.core.hits.connectHitsView
 import com.algolia.instantsearch.examples.directory.DirectoryItem
 import com.algolia.instantsearch.examples.directory.directoryItems
-import com.algolia.instantsearch.examples.directory.guides
+import com.algolia.instantsearch.examples.directory.navigateTo
 import com.algolia.instantsearch.examples.showcase.compose.client
 import com.algolia.instantsearch.examples.showcase.compose.configureSearchBox
 import com.algolia.instantsearch.examples.showcase.compose.ui.ShowcaseTheme
 import com.algolia.instantsearch.searcher.hits.HitsSearcher
 import com.algolia.search.model.IndexName
 import com.algolia.search.model.search.Query
-import com.algolia.search.serialize.KeyIndexName
-import com.algolia.search.serialize.KeyName
 
 class ComposeDirectoryShowcase : ComponentActivity() {
 
@@ -36,19 +33,11 @@ class ComposeDirectoryShowcase : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             ShowcaseTheme {
-                Directory(searchBoxState, hitsState, ::navigateToShowcase)
+                Directory(searchBoxState, hitsState, ::navigateTo)
             }
         }
         configureSearchBox(searcher, searchBoxState, connections)
         searcher.searchAsync()
-    }
-
-    private fun navigateToShowcase(item: DirectoryItem.Item) {
-        val intent = Intent(this, showcases.getValue(item.hit.objectID).java).apply {
-            putExtra(KeyIndexName, item.hit.index)
-            putExtra(KeyName, item.hit.name)
-        }
-        startActivity(intent)
     }
 
     override fun onDestroy() {
